+++ midgardmodels/v1/golang/issue.go
@@ -67,7 +67,7 @@ func (o IssuesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o IssuesList) Version() float64 {
+func (o IssuesList) Version() int {
 
 	return 1.0
 }
@@ -89,7 +89,7 @@ type Issue struct {
 	// Validity configures the max validity time for a token. If it is bigger than the configured max validity, it will be capped.
 	Validity string `json:"validity" bson:"validity"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -122,7 +122,7 @@ func (o *Issue) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *Issue) Version() float64 {
+func (o *Issue) Version() int {
 
 	return 1.0
 }
