+++ squallmodels/v1/golang/apiauthorizationpolicy.go
@@ -42,7 +42,7 @@ func (o APIAuthorizationPoliciesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o APIAuthorizationPoliciesList) Version() float64 {
+func (o APIAuthorizationPoliciesList) Version() int {
 
 	return 1.0
 }
@@ -100,7 +100,7 @@ type APIAuthorizationPolicy struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -137,7 +137,7 @@ func (o *APIAuthorizationPolicy) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *APIAuthorizationPolicy) Version() float64 {
+func (o *APIAuthorizationPolicy) Version() int {
 
 	return 1.0
 }
