+++ squallmodels/v1/golang/apicheck.go
@@ -64,7 +64,7 @@ func (o APIChecksList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o APIChecksList) Version() float64 {
+func (o APIChecksList) Version() int {
 
 	return 1.0
 }
@@ -86,7 +86,7 @@ type APICheck struct {
 	// Token is the token to use to check api authentication
 	Token string `json:"token" bson:"-"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -119,7 +119,7 @@ func (o *APICheck) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *APICheck) Version() float64 {
+func (o *APICheck) Version() int {
 
 	return 1.0
 }
