+++ squallmodels/v1/golang/dependencymap.go
@@ -40,7 +40,7 @@ func (o DependencyMapsList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o DependencyMapsList) Version() float64 {
+func (o DependencyMapsList) Version() int {
 
 	return 1.0
 }
@@ -59,7 +59,7 @@ type DependencyMap struct {
 	// nodes refers to the nodes of the map
 	Nodes gaiatypes.GraphNodeMap `json:"nodes" bson:"-"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -94,7 +94,7 @@ func (o *DependencyMap) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *DependencyMap) Version() float64 {
+func (o *DependencyMap) Version() int {
 
 	return 1.0
 }
