+++ squallmodels/v1/golang/enforcer.go
@@ -73,7 +73,7 @@ func (o EnforcersList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o EnforcersList) Version() float64 {
+func (o EnforcersList) Version() int {
 
 	return 1.0
 }
@@ -140,7 +140,7 @@ type Enforcer struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -178,7 +178,7 @@ func (o *Enforcer) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *Enforcer) Version() float64 {
+func (o *Enforcer) Version() int {
 
 	return 1.0
 }
