+++ squallmodels/v1/golang/enforcerprofile.go
@@ -53,7 +53,7 @@ func (o EnforcerProfilesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o EnforcerProfilesList) Version() float64 {
+func (o EnforcerProfilesList) Version() int {
 
 	return 1.0
 }
@@ -150,7 +150,7 @@ type EnforcerProfile struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -201,7 +201,7 @@ func (o *EnforcerProfile) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *EnforcerProfile) Version() float64 {
+func (o *EnforcerProfile) Version() int {
 
 	return 1.0
 }
