+++ squallmodels/v1/golang/fileaccess.go
@@ -52,7 +52,7 @@ func (o FileAccessList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o FileAccessList) Version() float64 {
+func (o FileAccessList) Version() int {
 
 	return 1.0
 }
@@ -77,7 +77,7 @@ type FileAccess struct {
 	// Protocol is the protocol used to access the file.
 	Protocol string `json:"protocol" bson:"-"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -108,7 +108,7 @@ func (o *FileAccess) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *FileAccess) Version() float64 {
+func (o *FileAccess) Version() int {
 
 	return 1.0
 }
