+++ squallmodels/v1/golang/fileaccesspolicy.go
@@ -42,7 +42,7 @@ func (o FileAccessPoliciesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o FileAccessPoliciesList) Version() float64 {
+func (o FileAccessPoliciesList) Version() int {
 
 	return 1.0
 }
@@ -112,7 +112,7 @@ type FileAccessPolicy struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -148,7 +148,7 @@ func (o *FileAccessPolicy) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *FileAccessPolicy) Version() float64 {
+func (o *FileAccessPolicy) Version() int {
 
 	return 1.0
 }
