+++ squallmodels/v1/golang/filepath.go
@@ -42,7 +42,7 @@ func (o FilePathsList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o FilePathsList) Version() float64 {
+func (o FilePathsList) Version() int {
 
 	return 1.0
 }
@@ -88,7 +88,7 @@ type FilePath struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -124,7 +124,7 @@ func (o *FilePath) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *FilePath) Version() float64 {
+func (o *FilePath) Version() int {
 
 	return 1.0
 }
