+++ squallmodels/v1/golang/flowstatistic.go
@@ -71,7 +71,7 @@ func (o FlowStatisticsList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o FlowStatisticsList) Version() float64 {
+func (o FlowStatisticsList) Version() int {
 
 	return 1.0
 }
@@ -108,7 +108,7 @@ type FlowStatistic struct {
 	// UserIdentifier can be set by the user as a query parameter. It will be returned in the FlowStatistic object.
 	UserIdentifier string `json:"userIdentifier" bson:"-"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -146,7 +146,7 @@ func (o *FlowStatistic) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *FlowStatistic) Version() float64 {
+func (o *FlowStatistic) Version() int {
 
 	return 1.0
 }
