+++ squallmodels/v1/golang/hookpolicy.go
@@ -56,7 +56,7 @@ func (o HookPoliciesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o HookPoliciesList) Version() float64 {
+func (o HookPoliciesList) Version() int {
 
 	return 1.0
 }
@@ -123,7 +123,7 @@ type HookPolicy struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -160,7 +160,7 @@ func (o *HookPolicy) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *HookPolicy) Version() float64 {
+func (o *HookPolicy) Version() int {
 
 	return 1.0
 }
