+++ squallmodels/v1/golang/namespace.go
@@ -40,7 +40,7 @@ func (o NamespacesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o NamespacesList) Version() float64 {
+func (o NamespacesList) Version() int {
 
 	return 1.0
 }
@@ -80,7 +80,7 @@ type Namespace struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -116,7 +116,7 @@ func (o *Namespace) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *Namespace) Version() float64 {
+func (o *Namespace) Version() int {
 
 	return 1.0
 }
