+++ squallmodels/v1/golang/quotapolicy.go
@@ -42,7 +42,7 @@ func (o QuotaPoliciesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o QuotaPoliciesList) Version() float64 {
+func (o QuotaPoliciesList) Version() int {
 
 	return 1.0
 }
@@ -100,7 +100,7 @@ type QuotaPolicy struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -136,7 +136,7 @@ func (o *QuotaPolicy) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *QuotaPolicy) Version() float64 {
+func (o *QuotaPolicy) Version() int {
 
 	return 1.0
 }
