+++ squallmodels/v1/golang/systemcall.go
@@ -42,7 +42,7 @@ func (o SystemCallsList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o SystemCallsList) Version() float64 {
+func (o SystemCallsList) Version() int {
 
 	return 1.0
 }
@@ -82,7 +82,7 @@ type SystemCall struct {
 	// UpdateTime is the time at which an entity was updated.
 	UpdateTime time.Time `json:"updateTime" bson:"updatetime"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -118,7 +118,7 @@ func (o *SystemCall) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *SystemCall) Version() float64 {
+func (o *SystemCall) Version() int {
 
 	return 1.0
 }
