+++ vincemodels/v1/golang/activate.go
@@ -38,7 +38,7 @@ func (o ActivatesList) DefaultOrder() []string {
 }
 
 // Version returns the version of the content.
-func (o ActivatesList) Version() float64 {
+func (o ActivatesList) Version() int {
 
 	return 1.0
 }
@@ -48,7 +48,7 @@ type Activate struct {
 	// Token contains the activation token
 	Token string `json:"token" bson:"-"`
 
-	ModelVersion float64 `json:"-" bson:"_modelversion"`
+	ModelVersion int `json:"-" bson:"_modelversion"`
 
 	sync.Mutex
 }
@@ -79,7 +79,7 @@ func (o *Activate) SetIdentifier(ID string) {
 }
 
 // Version returns the hardcoded version of the model
-func (o *Activate) Version() float64 {
+func (o *Activate) Version() int {
 
 	return 1.0
 }
